+++ boot_test.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/go-nacelle/service"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -12,7 +13,7 @@ func TestBoot(t *testing.T) {
 	ran := false
 	bootstrapper := NewBootstrapper(
 		"APP",
-		func(processes ProcessContainer, services ServiceContainer) error {
+		func(processes ProcessContainer, services *ServiceContainer) error {
 			processes.RegisterInitializer(InitializerFunc(func(ctx context.Context) error {
 				ran = true
 				return nil
@@ -28,15 +29,15 @@ func TestBoot(t *testing.T) {
 
 func TestDefaultServices(t *testing.T) {
 	serviceChecker := &struct {
-		Health   Health           `service:"health"`
-		Logger   Logger           `service:"logger"`
-		Services ServiceContainer `service:"services"`
+		Health   Health            `service:"health"`
+		Logger   Logger            `service:"logger"`
+		Services *ServiceContainer `service:"services"`
 	}{}
 
 	bootstrapper := NewBootstrapper(
 		"APP",
-		func(processes ProcessContainer, services ServiceContainer) error {
-			return services.Inject(serviceChecker)
+		func(processes ProcessContainer, services *ServiceContainer) error {
+			return service.Inject(services, serviceChecker)
 		},
 	)
 
@@ -49,7 +50,7 @@ func TestDefaultServices(t *testing.T) {
 func TestInitFuncError(t *testing.T) {
 	bootstrapper := NewBootstrapper(
 		"APP",
-		func(processes ProcessContainer, services ServiceContainer) error {
+		func(processes ProcessContainer, services *ServiceContainer) error {
 			return fmt.Errorf("oops")
 		},
 	)
@@ -60,7 +61,7 @@ func TestInitFuncError(t *testing.T) {
 func TestLoggingInitError(t *testing.T) {
 	bootstrapper := NewBootstrapper(
 		"APP",
-		func(processes ProcessContainer, services ServiceContainer) error {
+		func(processes ProcessContainer, services *ServiceContainer) error {
 			return nil
 		},
 		WithLoggingInitFunc(func(*Config) (Logger, error) {
@@ -74,7 +75,7 @@ func TestLoggingInitError(t *testing.T) {
 func TestRunnerError(t *testing.T) {
 	bootstrapper := NewBootstrapper(
 		"APP",
-		func(processes ProcessContainer, services ServiceContainer) error {
+		func(processes ProcessContainer, services *ServiceContainer) error {
 			processes.RegisterInitializer(InitializerFunc(func(ctx context.Context) error {
 				return fmt.Errorf("oops")
 			}))