+++ service_imports.go
@@ -1,10 +1,9 @@
 package nacelle
 
-import "github.com/go-nacelle/service"
-
-type ServiceContainer = service.ServiceContainer
-
-var (
-	NewServiceContainer = service.NewServiceContainer
-	Overlay             = service.Overlay
+import (
+	"github.com/go-nacelle/service"
 )
+
+type ServiceContainer = service.Container
+
+var NewServiceContainer = service.New