+++ CHANGELOG.md
@@ -13,10 +13,12 @@
 ### Changed
 
 - Changed signature of `ServiceInitializerFunc`. [#11](https://github.com/go-nacelle/nacelle/pull/11)
+- Changed signature of `AppInitFunc`, `ServiceInitializerFunc`, and `WrapServiceInitializerFunc`. [#17](https://github.com/go-nacelle/nacelle/pull/17)
 
 ### Removed
 
 - Removed mocks package. [#14](https://github.com/go-nacelle/nacelle/pull/14)
+- Removed `Overlay` import. [#17](https://github.com/go-nacelle/nacelle/pull/17)
 
 ## [v1.2.0] - 2020-10-04